权限创建与校验服务接口实现

+++ service/permission-service/src/main/java/cn/zeroclian/github/dao/UserRoleMappingRepository.java
@@ -15,5 +15,5 @@ public interface UserRoleMappingRepository extends JpaRepository<UserRoleMapping
      * @param userId    用户id
      * @return
      */
-    UserRoleMapping findByUserId(Integer userId);
+    UserRoleMapping findByUserId(Long userId);
 }
